refactor(k8s): decompose kubernetes.py into focused mixin modules
Split the 3,535-line kubernetes.py into 6 focused files using the mixin
pattern (same approach as existing KubernetesLauncherMixin):

- kubernetes.py (779 lines): core lifecycle (init, validate, prepare, deploy, monitor, cleanup)
- k8s_results.py (1,390 lines): results collection, PVC artifacts, perf CSV reporting
- k8s_template_context.py (856 lines): Jinja2 context, env vars, data config, tools
- k8s_scripts.py (352 lines): script/tool loading for ConfigMap embedding
- k8s_pvc.py (239 lines): PVC lifecycle management

All 426 unit tests and 139 integration tests pass unchanged.
